requirements.txt
python-telegram-bot==13.11
discord-webhook==0.15.0